\name{ICSOutlier-package}
\alias{ICSOutlier-package}
\alias{ICSOutlier}

\docType{package}

\title{
\packageTitle{ICSOutlier}
}
\description{
\packageDescription{ICSOutlier}
}
\details{
\packageDESCRIPTION{ICSOutlier}

ICS is a general multivariate technique with many applications in multivariate analysis. ICSOutlier
offers a selection of functions for automated detection of outliers in the data based on a fitted ics2
object. The current implementation targets data sets with only a small percentage of outliers but future extensions
are under preparation. 

\packageIndices{ICSOutlier}
}

\references{
\cite{Archimbaud, A., Nordhausen, K. and Ruiz-Gazen, A. (2018), ICS for multivariate outlier detection with application to quality control. Computational Statistics & Data Analysis, 128:184-199.  <doi:10.1016/j.csda.2018.06.011>.}   

\cite{Archimbaud, A., Nordhausen, K. and Ruiz-Gazen, A. (2018), ICSOutlier: Unsupervised Outlier Detection for Low-Dimensional Contamination Structure. The R Journal, 10:234-250.  <doi:10.32614/RJ-2018-034>.}   
}

\author{
\packageAuthor{ICSOutlier}

Maintainer: \packageMaintainer{ICSOutlier}
}

\keyword{package}


