% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ics.outlier.R
\name{print.ICS_Out}
\alias{print.ICS_Out}
\title{Vector of Outlier Indicators}
\usage{
\method{print}{ICS_Out}(x, ...)
}
\arguments{
\item{x}{object object of class \code{"ICS_Out"}.}

\item{...}{additional arguments, not used.}
}
\value{
The supplied object of class \code{"ICS_Out_summary"} is returned invisibly.
}
\description{
Short statement about how many components are selected for the outlier detection and how many outliers are detected.
}
\author{
Aurore Archimbaud and Klaus Nordhausen
}
