\name{print.icsOut}
\docType{methods}
\alias{print.icsOut}
%\alias{show-method}
\alias{show,icsOut-method}
\title{Vector of Outlier Indicators}
\description{
  Short statement about how many components are selected for the outlier detection and how many outliers are detected.
  
}
\usage{
\S4method{show}{icsOut}(object)
}

\arguments{
  \item{object}{object of class \code{icsOut}. }
}

\author{Aurore Archimbaud and Klaus Nordhausen}

\seealso{\code{\link{icsOut-class}} and \code{\link{ics.outlier}}}

\keyword{methods}
\keyword{print}

