% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ics.outlier.R
\name{print.ICSOut_summary}
\alias{print.ICSOut_summary}
\title{Print of an \code{ICSOut_summary} object}
\usage{
\method{print}{ICSOut_summary}(x, ...)
}
\arguments{
\item{x}{object of class \code{"ICSOut_summary"}.}

\item{...}{additional arguments, not used.}
}
\value{
The supplied object of class \code{"ICSOut_summary"} is returned invisibly.
}
\description{
Prints an \code{ICSOut_summary} object in an informative way.
}
\author{
Aurore Archimbaud and Klaus Nordhausen
}
