% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ics.outlier.R
\name{summary.ICS_Out}
\alias{summary.ICS_Out}
\title{Summary of an 'ICS_Out' Object

Summarizes an 'ICS_Out' object in an informative way.}
\usage{
\method{summary}{ICS_Out}(object, ...)
}
\arguments{
\item{object}{object object of class \code{"ICS_Out"}.}

\item{...}{additional arguments passed to \code{\link[=summary]{summary()}}}
}
\value{
An object of class \code{"ICS_Out_summary"} with the following components:
\itemize{
\item \code{comps}: Vector giving the indices of the ICS components selected.
\item \code{method}: Name of the method used to decide upon the number of ICS components.
\item \code{test}: he name of the normality test as specified in the function call.
\item \code{S1_label}:  Name of S1.
\item \code{S2_label}:  Name of S2.
\item \code{level_test}: The level for deciding upon the cut-off value for the ICS distances.
\item \code{level_dist}: The initial level for selecting the invariant coordinates.
\item \code{nb_outliers}: the number of observations identified as outliers.
}
}
\description{
Summary of an 'ICS_Out' Object

Summarizes an 'ICS_Out' object in an informative way.
}
\author{
Aurore Archimbaud and Klaus Nordhausen
}
