\name{summary.icsOut}
\docType{methods}
\alias{summary.icsOut}

\alias{summary,icsOut-method}

\title{Summarize a icsOut object}
\description{
  Summarizes and prints an \code{icsOut} object in an informative way.
}
\usage{
\S4method{summary}{icsOut}(object, digits = 4)
}

\arguments{
  \item{object}{object of class \code{icsOut}.}
  \item{digits}{ number of digits for the numeric output.}
}


\author{Aurore Archimbaud and Klaus Nordhausen}

\seealso{\code{\link{icsOut-class}} and \code{\link{ics.outlier}}}

\keyword{methods}
\keyword{print}
