% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HTP2}
\alias{HTP2}
\title{Production Measurements of High-Tech Parts}
\format{
A data frame with 457 rows and 149 variables V.1 - V.149, presenting
some collinearity issues.
}
\source{
Anonymized data from a nondisclosed manufacturer.
}
\usage{
data("HTP2")
}
\description{
The \code{HTP2} data set contains 457 high-tech parts designed for consumer
products characterized by 149 tests.
These tests are performed to ensure a high quality of the production.
All these 457 parts were considered functional and have been sold.
However the part 28 showed defects in use and was
returned to the manufacturer by the customer. Therefore this part
can be considered as outlier.
}
\keyword{datasets}
