\name{ICSOutlier}
\alias{ICSOutlier}

\docType{package}
\title{
Outlier Detection Using Invariant Coordinate Selection
}
\description{
Multivariate outlier detection is 
performed using invariant coordinates and the package 
offers different methods to choose the appropriate components. 
}
\details{
\tabular{ll}{
Package: \tab ICS\cr
Type: \tab Package\cr
Version: \tab 0.2-0\cr
Date: \tab 2016-12-20\cr
License: \tab GPL (>= 2)\cr
}

ICS is a general multivariate technique with many applications in multivariate analysis. ICSOutlier
offers a selection of functions for automated detection of outliers in the data based on a fitted ics2
object. The current implementation targets data sets with only a small percentage of outliers but future extensions
are under preparation. 
}


\author{
Aurore Archimbaud, Klaus Nordhausen, Anne Ruiz-Gazen

Maintainer: 
Klaus Nordhausen, \email{klaus.nordhausen@utu.fi}
}

\keyword{package}
