% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSOutlier-package.R
\docType{package}
\name{ICSOutlier-package}
\alias{ICSOutlier-package}
\alias{ICSOutlier}
\title{\packageTitle{ICSOutlier}}
\description{
\packageDescription{ICSOutlier}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{ICSOutlier}
\packageIndices{ICSOutlier}
}
\references{
Archimbaud, A., Nordhausen, K. and Ruiz-Gazen, A. (2018), ICS for multivariate outlier detection with application to quality control. Computational Statistics & Data Analysis, 128:184-199. ISSN 0167-9473.  \doi{10.1016/j.csda.2018.06.011}.

Archimbaud, A., Nordhausen, K. and Ruiz-Gazen, A. (2018), ICSOutlier: Unsupervised Outlier Detection for Low-Dimensional Contamination Structure. The R Journal, 10:234-250.  \doi{10.32614/RJ-2018-034}.
}
\author{
\packageAuthor{ICSOutlier}

Maintainer: \packageMaintainer{ICSOutlier}
}
\keyword{package}
